fix: change wind deg to double

+++ app/src/main/java/com/onadasoft/weatherdaily/models/Wind.java
@@ -2,7 +2,7 @@ package com.onadasoft.weatherdaily.models;
 
 public class Wind {
     private double speed;
-    private long deg;
+    private double deg;
 
     public double getSpeed() {
         return speed;
@@ -12,11 +12,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public long getDeg() {
+    public double getDeg() {
         return deg;
     }
 
-    public void setDeg(long deg) {
+    public void setDeg(double deg) {
         this.deg = deg;
     }
